requirements.txt
asgiref==3.2.7
astroid==2.4.1
Django==3.0.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
wrapt==1.12.1
psycopg2==2.8.5
Pillow==7.1.2
djangorestframework==3.11.0
django-filter==2.2.0
django-phone-verify==2.0.0
django-phone-verify==2.0.0
djangorestframework-simplejwt==4.4.0
social-auth-app-django==3.1.0
